pyproject.toml
[project]
name = "morph-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "starlette>=0.37.2",
    "uvicorn[standard]>=0.30.1",
    "PyGithub>=2.3.0",
    "claude-code-sdk>=0.0.14",
    "PyJWT>=2.8.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9"
]
